+++ src/lib/wapi/serializers/serielize-me.js
@@ -2,17 +2,22 @@ export const _serializeMeObj = async (obj) => {
   if (obj == undefined) {
     return null;
   }
+
+  let appsImage = undefined;
   const status = await Store.MyStatus.getStatus(obj.wid._serialized);
-  const image = await Store.Profile.profilePicFindThumbFromPhone(
-    obj.wid._serialized
-  );
-  const appsImage = Store.Profile.profilePicFind(obj.wid._serialized);
+  await Store.Profile.profilePicFind(obj.wid._serialized)
+    .then((e) => {
+      appsImage = e;
+    })
+    .catch(() => {});
+  const connection = await window.Store.State.default.state;
+
   return Object.assign(
     {},
     {
       battery: obj.battery,
-      connected: obj.connected,
       locales: obj.locales,
+      statusConnection: connection,
       phone: {
         device_manufacturer: obj.phone.device_manufacturer,
         device_model: obj.phone.device_model,
@@ -21,8 +26,7 @@ export const _serializeMeObj = async (obj) => {
       },
       pushname: obj.pushname,
       status: status.status,
-      imageBase64: image.raw,
-      appsImage: appsImage.eurl,
+      appsImage: appsImage && appsImage.eurl ? appsImage.eurl : undefined,
       wid: {
         server: obj.wid.server,
         user: obj.wid.user,