+++ src/lib/wapi/wapi.js
@@ -63,7 +63,8 @@ import {
   clearChatMessages,
   createGroup,
   deleteConversation,
-  deleteMessages,
+  deleteMessagesAll,
+  deleteMessagesMe,
   demoteParticipant,
   downloadFile,
   encryptAndUploadFile,
@@ -216,10 +217,11 @@ var loadParasite = function () {
     try {
       const last = window['webpackChunkwhatsapp_web_client'].length - 1;
       if (
-        !/^parasite/.test(
-          window['webpackChunkwhatsapp_web_client'][last][0][0]
+        !window['webpackChunkwhatsapp_web_client'][last][0].includes(
+          'parasite'
         ) &&
-        (document.querySelectorAll('#app .two').length ||
+        (document.querySelectorAll('#app').length ||
+          document.querySelectorAll('#app .two').length ||
           document.querySelector('canvas') ||
           document.querySelectorAll('#startup').length == 0)
       ) {
@@ -239,7 +241,6 @@ if (typeof window.WAPI === 'undefined') {
   window.WAPI.interfaceMute = interfaceMute;
   window.WAPI.checkIdMessage = checkIdMessage;
   window.WAPI.returnReply = returnReply;
-  window.WAPI.getStore = getStore;
   window.WAPI.checkChat = checkChat;
   window.WAPI.checkNumberStatus = checkNumberStatus;
   window.WAPI.sendCheckType = sendCheckType;
@@ -293,7 +294,8 @@ if (typeof window.WAPI === 'undefined') {
   window.WAPI.sendMessage2 = sendMessage2;
   window.WAPI.sendSeen = sendSeen;
   window.WAPI.deleteConversation = deleteConversation;
-  window.WAPI.deleteMessages = deleteMessages;
+  window.WAPI.deleteMessagesAll = deleteMessagesAll;
+  window.WAPI.deleteMessagesMe = deleteMessagesMe;
   window.WAPI.clearChatMessages = clearChatMessages;
   window.WAPI.sendImage = sendImage;
   window.WAPI.sendPtt = sendPtt;